(REF) Extract BaseApplication
