+++ bin/cv
@@ -26,4 +26,4 @@ foreach ($autoloaders as $autoloader) {
 if (!$found) {
   die("Failed to find autoloader");
 }
-\Civi\Cv\Application::main(__DIR__, $argv);
+\Civi\Cv\Application::main('cv', __DIR__, $argv);
